make alternate locations independent
Kept the original metadata then added three new items. Two of them are like the rest where it is an array of strting that all index with the FileName. They are the compression algorithm (default is none) and if the reference is available.

The third is now a list of a list of objects that represent the alternate locations. When they are added, all of the same metadata other than alternate locations is expected.

Since harvest is ingesting what it fines, it populates the original allowing for all of the others fields to be populated when the alternate location is known.