fix latest lidvid
There were three messes contributing to these problems. mixing lids and lidvids, counting on string sorting, and bad aggregate.

Simple typo was mising lids and lidvids when making the list. Hence, everything was just getting globbed together and it looked like all 337 items were updating. Counted correctly but counted wrong list.

The aggregate was asking opensearch to sort the results into a bucket. It was. However it was sorting the entire list of lids onto the same bucket so always getting the top versions of just one lid. Changed the aggregate to create a bucket for each lid. Since string sorting the version is just wrong, fixed that before somebody found it as a bug.

Instead of asking opensearch to sort, do it locally with the understanding that the VID is major.minor and both are integers. Works jusst fine.